[#159] iOS17 isSignedIn 버그 픽스
